development/server/algorithm/tf_faster_rcnn/requirements.txt
absl-py==0.4.1
astor==0.7.1
certifi==2018.8.24
cloudpickle==0.5.6
cycler==0.10.0
Cython==0.28.5
dask==0.19.2
decorator==4.3.0
easydict==1.8
gast==0.2.0
grpcio==1.12.1
kiwisolver==1.0.1
lxml==4.2.5
Markdown==2.6.11
matplotlib==3.0.0
mkl-fft==1.0.4
mkl-random==1.0.1
networkx==2.2
numpy==1.15.1
opencv-python==3.4.3.18
Pillow==5.2.0
prettytable==0.7.2
protobuf==3.6.0
pycocotools==2.0.0
pyparsing==2.2.1
python-dateutil==2.7.3
PyWavelets==1.0.0
PyYAML==3.13
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1
